requirements.txt
django
djangorestframework
pillow
psycopg2-binary
python-dotenv
drf-spectacular
django-jazzmin
redis
python-decouple
celery
django_celery_beat